refactor: rename existing legacy token files
