+++ packages/eui/src-docs/src/views/theme/_json/eui_theme_amsterdam_dark.json
