+++ packages/eui/src-docs/src/views/theme/typography/values.tsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { css } from '@emotion/react';
 
 // @ts-ignore Importing JSON
-import themeVars from '../_json/eui_theme_light.json'; // Only grabbing font weight vars, so color mode is irrelevant
+import themeVars from '../_json/eui_theme_amsterdam_light.json'; // Only grabbing font weight vars, so color mode is irrelevant
 
 import { GuideSection } from '../../../components/guide_section/guide_section';
